rewritten tests to single fixture with patches
