+++ tests/components/wallbox/test_number.py
@@ -1,29 +1,22 @@
 """Test Wallbox Switch component."""
 
-from unittest.mock import Mock, patch
+from unittest.mock import patch
 
 import pytest
 
 from homeassistant.components.input_number import ATTR_VALUE, SERVICE_SET_VALUE
 from homeassistant.components.number import DOMAIN as NUMBER_DOMAIN
-from homeassistant.components.wallbox.const import (
-    CHARGER_ENERGY_PRICE_KEY,
-    CHARGER_MAX_CHARGING_CURRENT_KEY,
-    CHARGER_MAX_ICP_CURRENT_KEY,
-)
 from homeassistant.components.wallbox.coordinator import InvalidAuth
 from homeassistant.const import ATTR_ENTITY_ID
 from homeassistant.core import HomeAssistant
 from homeassistant.exceptions import HomeAssistantError
 
-from . import (
-    authorisation_response,
+from .conftest import (
     http_403_error,
     http_404_error,
     http_429_error,
     setup_integration,
-    setup_integration_bidir,
-    setup_integration_platform_not_ready,
+    test_response_bidir,
 )
 from .const import (
     MOCK_NUMBER_ENTITY_ENERGY_PRICE_ID,
@@ -33,103 +26,69 @@ from .const import (
 
 from tests.common import MockConfigEntry
 
-mock_wallbox = Mock()
-mock_wallbox.authenticate = Mock(return_value=authorisation_response)
-mock_wallbox.setEnergyCost = Mock(return_value={CHARGER_ENERGY_PRICE_KEY: 1.1})
-mock_wallbox.setMaxChargingCurrent = Mock(
-    return_value={CHARGER_MAX_CHARGING_CURRENT_KEY: 20}
-)
-mock_wallbox.setIcpMaxCurrent = Mock(return_value={CHARGER_MAX_ICP_CURRENT_KEY: 10})
-
 
 async def test_wallbox_number_class(
-    hass: HomeAssistant, entry: MockConfigEntry
+    hass: HomeAssistant, entry: MockConfigEntry, mock_wallbox
 ) -> None:
     """Test wallbox sensor class."""
-
     await setup_integration(hass, entry)
 
-    with (
-        patch(
-            "homeassistant.components.wallbox.Wallbox.authenticate",
-            new=Mock(return_value=authorisation_response),
-        ),
-        patch(
-            "homeassistant.components.wallbox.Wallbox.setMaxChargingCurrent",
-            new=Mock(return_value={CHARGER_MAX_CHARGING_CURRENT_KEY: 20}),
-        ),
-    ):
-        state = hass.states.get(MOCK_NUMBER_ENTITY_ID)
-        assert state.attributes["min"] == 6
-        assert state.attributes["max"] == 25
+    state = hass.states.get(MOCK_NUMBER_ENTITY_ID)
+    assert state.attributes["min"] == 6
+    assert state.attributes["max"] == 25
 
-        await hass.services.async_call(
-            "number",
-            SERVICE_SET_VALUE,
-            {
-                ATTR_ENTITY_ID: MOCK_NUMBER_ENTITY_ID,
-                ATTR_VALUE: 20,
-            },
-            blocking=True,
-        )
+    await hass.services.async_call(
+        "number",
+        SERVICE_SET_VALUE,
+        {
+            ATTR_ENTITY_ID: MOCK_NUMBER_ENTITY_ID,
+            ATTR_VALUE: 20,
+        },
+        blocking=True,
+    )
 
 
 async def test_wallbox_number_class_bidir(
-    hass: HomeAssistant, entry: MockConfigEntry
+    hass: HomeAssistant, entry: MockConfigEntry, mock_wallbox
 ) -> None:
     """Test wallbox sensor class."""
+    with patch.object(
+        mock_wallbox, "getChargerStatus", return_value=test_response_bidir
+    ):
+        await setup_integration(hass, entry)
 
-    await setup_integration_bidir(hass, entry)
-
-    state = hass.states.get(MOCK_NUMBER_ENTITY_ID)
-    assert state.attributes["min"] == -25
-    assert state.attributes["max"] == 25
+        state = hass.states.get(MOCK_NUMBER_ENTITY_ID)
+        assert state.attributes["min"] == -25
+        assert state.attributes["max"] == 25
 
 
 async def test_wallbox_number_energy_class(
-    hass: HomeAssistant, entry: MockConfigEntry
+    hass: HomeAssistant, entry: MockConfigEntry, mock_wallbox
 ) -> None:
     """Test wallbox sensor class."""
 
     await setup_integration(hass, entry)
 
-    with (
-        patch(
-            "homeassistant.components.wallbox.Wallbox.authenticate",
-            new=Mock(return_value=authorisation_response),
-        ),
-        patch(
-            "homeassistant.components.wallbox.Wallbox.setEnergyCost",
-            new=Mock(return_value={CHARGER_ENERGY_PRICE_KEY: 1.1}),
-        ),
-    ):
-        await hass.services.async_call(
-            "number",
-            SERVICE_SET_VALUE,
-            {
-                ATTR_ENTITY_ID: MOCK_NUMBER_ENTITY_ENERGY_PRICE_ID,
-                ATTR_VALUE: 1.1,
-            },
-            blocking=True,
-        )
+    await hass.services.async_call(
+        "number",
+        SERVICE_SET_VALUE,
+        {
+            ATTR_ENTITY_ID: MOCK_NUMBER_ENTITY_ENERGY_PRICE_ID,
+            ATTR_VALUE: 1.1,
+        },
+        blocking=True,
+    )
 
 
 async def test_wallbox_number_class_connection_error(
-    hass: HomeAssistant, entry: MockConfigEntry
+    hass: HomeAssistant, entry: MockConfigEntry, mock_wallbox
 ) -> None:
     """Test wallbox sensor class."""
 
     await setup_integration(hass, entry)
 
     with (
-        patch(
-            "homeassistant.components.wallbox.Wallbox.authenticate",
-            new=Mock(return_value=authorisation_response),
-        ),
-        patch(
-            "homeassistant.components.wallbox.Wallbox.setMaxChargingCurrent",
-            new=Mock(side_effect=http_404_error),
-        ),
+        patch.object(mock_wallbox, "setMaxChargingCurrent", side_effect=http_404_error),
         pytest.raises(HomeAssistantError),
     ):
         await hass.services.async_call(
@@ -144,21 +103,14 @@ async def test_wallbox_number_class_connection_error(
 
 
 async def test_wallbox_number_class_too_many_requests(
-    hass: HomeAssistant, entry: MockConfigEntry
+    hass: HomeAssistant, entry: MockConfigEntry, mock_wallbox
 ) -> None:
     """Test wallbox sensor class."""
 
     await setup_integration(hass, entry)
 
     with (
-        patch(
-            "homeassistant.components.wallbox.Wallbox.authenticate",
-            new=Mock(return_value=authorisation_response),
-        ),
-        patch(
-            "homeassistant.components.wallbox.Wallbox.setMaxChargingCurrent",
-            new=Mock(side_effect=http_429_error),
-        ),
+        patch.object(mock_wallbox, "setMaxChargingCurrent", side_effect=http_429_error),
         pytest.raises(HomeAssistantError),
     ):
         await hass.services.async_call(
@@ -173,21 +125,14 @@ async def test_wallbox_number_class_too_many_requests(
 
 
 async def test_wallbox_number_class_energy_price_update_failed(
-    hass: HomeAssistant, entry: MockConfigEntry
+    hass: HomeAssistant, entry: MockConfigEntry, mock_wallbox
 ) -> None:
     """Test wallbox sensor class."""
 
     await setup_integration(hass, entry)
 
     with (
-        patch(
-            "homeassistant.components.wallbox.Wallbox.authenticate",
-            new=Mock(return_value=authorisation_response),
-        ),
-        patch(
-            "homeassistant.components.wallbox.Wallbox.setEnergyCost",
-            new=Mock(side_effect=http_429_error),
-        ),
+        patch.object(mock_wallbox, "setEnergyCost", side_effect=http_429_error),
         pytest.raises(HomeAssistantError),
     ):
         await hass.services.async_call(
@@ -202,21 +147,14 @@ async def test_wallbox_number_class_energy_price_update_failed(
 
 
 async def test_wallbox_number_class_energy_price_update_connection_error(
-    hass: HomeAssistant, entry: MockConfigEntry
+    hass: HomeAssistant, entry: MockConfigEntry, mock_wallbox
 ) -> None:
     """Test wallbox sensor class."""
 
     await setup_integration(hass, entry)
 
     with (
-        patch(
-            "homeassistant.components.wallbox.Wallbox.authenticate",
-            new=Mock(return_value=authorisation_response),
-        ),
-        patch(
-            "homeassistant.components.wallbox.Wallbox.setEnergyCost",
-            new=Mock(side_effect=http_404_error),
-        ),
+        patch.object(mock_wallbox, "setEnergyCost", side_effect=http_404_error),
         pytest.raises(HomeAssistantError),
     ):
         await hass.services.async_call(
@@ -231,21 +169,14 @@ async def test_wallbox_number_class_energy_price_update_connection_error(
 
 
 async def test_wallbox_number_class_energy_price_auth_error(
-    hass: HomeAssistant, entry: MockConfigEntry
+    hass: HomeAssistant, entry: MockConfigEntry, mock_wallbox
 ) -> None:
     """Test wallbox sensor class."""
 
     await setup_integration(hass, entry)
 
     with (
-        patch(
-            "homeassistant.components.wallbox.Wallbox.authenticate",
-            new=Mock(return_value=authorisation_response),
-        ),
-        patch(
-            "homeassistant.components.wallbox.Wallbox.setEnergyCost",
-            new=Mock(side_effect=http_429_error),
-        ),
+        patch.object(mock_wallbox, "setEnergyCost", side_effect=http_429_error),
         pytest.raises(HomeAssistantError),
     ):
         await hass.services.async_call(
@@ -259,40 +190,40 @@ async def test_wallbox_number_class_energy_price_auth_error(
         )
 
 
-async def test_wallbox_number_class_platform_not_ready(
-    hass: HomeAssistant, entry: MockConfigEntry
-) -> None:
-    """Test wallbox lock not loaded on authentication error."""
-
-    await setup_integration_platform_not_ready(hass, entry)
-
-    state = hass.states.get(MOCK_NUMBER_ENTITY_ID)
-
-    assert state is None
-
-
 async def test_wallbox_number_class_icp_energy(
-    hass: HomeAssistant, entry: MockConfigEntry
+    hass: HomeAssistant, entry: MockConfigEntry, mock_wallbox
+) -> None:
+    """Test wallbox sensor class."""
+
+    await setup_integration(hass, entry)
+
+    await hass.services.async_call(
+        NUMBER_DOMAIN,
+        SERVICE_SET_VALUE,
+        {
+            ATTR_ENTITY_ID: MOCK_NUMBER_ENTITY_ICP_CURRENT_ID,
+            ATTR_VALUE: 10,
+        },
+        blocking=True,
+    )
+
+
+async def test_wallbox_number_class_max_current_auth_error(
+    hass: HomeAssistant, entry: MockConfigEntry, mock_wallbox
 ) -> None:
     """Test wallbox sensor class."""
 
     await setup_integration(hass, entry)
 
     with (
-        patch(
-            "homeassistant.components.wallbox.Wallbox.authenticate",
-            new=Mock(return_value=authorisation_response),
-        ),
-        patch(
-            "homeassistant.components.wallbox.Wallbox.setIcpMaxCurrent",
-            new=Mock(return_value={CHARGER_MAX_ICP_CURRENT_KEY: 10}),
-        ),
+        patch.object(mock_wallbox, "setMaxChargingCurrent", side_effect=http_403_error),
+        pytest.raises(InvalidAuth),
     ):
         await hass.services.async_call(
             NUMBER_DOMAIN,
             SERVICE_SET_VALUE,
             {
-                ATTR_ENTITY_ID: MOCK_NUMBER_ENTITY_ICP_CURRENT_ID,
+                ATTR_ENTITY_ID: MOCK_NUMBER_ENTITY_ID,
                 ATTR_VALUE: 10,
             },
             blocking=True,
@@ -300,21 +231,14 @@ async def test_wallbox_number_class_icp_energy(
 
 
 async def test_wallbox_number_class_icp_energy_auth_error(
-    hass: HomeAssistant, entry: MockConfigEntry
+    hass: HomeAssistant, entry: MockConfigEntry, mock_wallbox
 ) -> None:
     """Test wallbox sensor class."""
 
     await setup_integration(hass, entry)
 
     with (
-        patch(
-            "homeassistant.components.wallbox.Wallbox.authenticate",
-            new=Mock(return_value=authorisation_response),
-        ),
-        patch(
-            "homeassistant.components.wallbox.Wallbox.setIcpMaxCurrent",
-            new=Mock(side_effect=http_403_error),
-        ),
+        patch.object(mock_wallbox, "setIcpMaxCurrent", side_effect=http_403_error),
         pytest.raises(InvalidAuth),
     ):
         await hass.services.async_call(
@@ -329,21 +253,14 @@ async def test_wallbox_number_class_icp_energy_auth_error(
 
 
 async def test_wallbox_number_class_icp_energy_connection_error(
-    hass: HomeAssistant, entry: MockConfigEntry
+    hass: HomeAssistant, entry: MockConfigEntry, mock_wallbox
 ) -> None:
     """Test wallbox sensor class."""
 
     await setup_integration(hass, entry)
 
     with (
-        patch(
-            "homeassistant.components.wallbox.Wallbox.authenticate",
-            new=Mock(return_value=authorisation_response),
-        ),
-        patch(
-            "homeassistant.components.wallbox.Wallbox.setIcpMaxCurrent",
-            new=Mock(side_effect=http_404_error),
-        ),
+        patch.object(mock_wallbox, "setIcpMaxCurrent", side_effect=http_404_error),
         pytest.raises(HomeAssistantError),
     ):
         await hass.services.async_call(
@@ -358,21 +275,14 @@ async def test_wallbox_number_class_icp_energy_connection_error(
 
 
 async def test_wallbox_number_class_icp_energy_too_many_request(
-    hass: HomeAssistant, entry: MockConfigEntry
+    hass: HomeAssistant, entry: MockConfigEntry, mock_wallbox
 ) -> None:
     """Test wallbox sensor class."""
 
     await setup_integration(hass, entry)
 
     with (
-        patch(
-            "homeassistant.components.wallbox.Wallbox.authenticate",
-            new=Mock(return_value=authorisation_response),
-        ),
-        patch(
-            "homeassistant.components.wallbox.Wallbox.setIcpMaxCurrent",
-            new=Mock(side_effect=http_429_error),
-        ),
+        patch.object(mock_wallbox, "setIcpMaxCurrent", side_effect=http_429_error),
         pytest.raises(HomeAssistantError),
     ):
         await hass.services.async_call(
