+++ tests/components/wallbox/test_select.py
@@ -1,6 +1,6 @@
 """Test Wallbox Select component."""
 
-from unittest.mock import Mock, patch
+from unittest.mock import patch
 
 import pytest
 
@@ -9,15 +9,14 @@ from homeassistant.components.select import (
     DOMAIN as SELECT_DOMAIN,
     SERVICE_SELECT_OPTION,
 )
-from homeassistant.components.wallbox.const import CHARGER_STATUS_ID_KEY, EcoSmartMode
+from homeassistant.components.wallbox.const import EcoSmartMode
 from homeassistant.const import ATTR_ENTITY_ID
 from homeassistant.core import HomeAssistant, HomeAssistantError
 
-from . import (
-    authorisation_response,
+from .conftest import (
     http_404_error,
     http_429_error,
-    setup_integration_select,
+    setup_integration,
     test_response,
     test_response_eco_mode,
     test_response_full_solar,
@@ -34,33 +33,13 @@ TEST_OPTIONS = [
 ]
 
 
-@pytest.fixture
-def mock_authenticate():
-    """Fixture to patch Wallbox methods."""
-    with patch(
-        "homeassistant.components.wallbox.Wallbox.authenticate",
-        new=Mock(return_value=authorisation_response),
-    ):
-        yield
-
-
 @pytest.mark.parametrize(("mode", "response"), TEST_OPTIONS)
 async def test_wallbox_select_solar_charging_class(
-    hass: HomeAssistant, entry: MockConfigEntry, mode, response, mock_authenticate
+    hass: HomeAssistant, entry: MockConfigEntry, mode, response, mock_wallbox
 ) -> None:
     """Test wallbox select class."""
-
-    with (
-        patch(
-            "homeassistant.components.wallbox.Wallbox.enableEcoSmart",
-            new=Mock(return_value={CHARGER_STATUS_ID_KEY: 193}),
-        ),
-        patch(
-            "homeassistant.components.wallbox.Wallbox.disableEcoSmart",
-            new=Mock(return_value={CHARGER_STATUS_ID_KEY: 193}),
-        ),
-    ):
-        await setup_integration_select(hass, entry, response)
+    with patch.object(mock_wallbox, "getChargerStatus", return_value=response):
+        await setup_integration(hass, entry)
 
         await hass.services.async_call(
             SELECT_DOMAIN,
@@ -77,39 +56,35 @@ async def test_wallbox_select_solar_charging_class(
 
 
 async def test_wallbox_select_no_power_boost_class(
-    hass: HomeAssistant, entry: MockConfigEntry
+    hass: HomeAssistant, entry: MockConfigEntry, mock_wallbox
 ) -> None:
     """Test wallbox select class."""
 
-    await setup_integration_select(hass, entry, test_response_no_power_boost)
+    with patch.object(
+        mock_wallbox, "getChargerStatus", return_value=test_response_no_power_boost
+    ):
+        await setup_integration(hass, entry)
 
-    state = hass.states.get(MOCK_SELECT_ENTITY_ID)
-    assert state is None
+        state = hass.states.get(MOCK_SELECT_ENTITY_ID)
+        assert state is None
 
 
 @pytest.mark.parametrize(("mode", "response"), TEST_OPTIONS)
-@pytest.mark.parametrize("error", [http_404_error, ConnectionError])
 async def test_wallbox_select_class_error(
     hass: HomeAssistant,
     entry: MockConfigEntry,
     mode,
     response,
-    error,
-    mock_authenticate,
+    mock_wallbox,
 ) -> None:
     """Test wallbox select class connection error."""
 
-    await setup_integration_select(hass, entry, response)
+    await setup_integration(hass, entry)
 
     with (
-        patch(
-            "homeassistant.components.wallbox.Wallbox.disableEcoSmart",
-            new=Mock(side_effect=error),
-        ),
-        patch(
-            "homeassistant.components.wallbox.Wallbox.enableEcoSmart",
-            new=Mock(side_effect=error),
-        ),
+        patch.object(mock_wallbox, "getChargerStatus", return_value=response),
+        patch.object(mock_wallbox, "disableEcoSmart", side_effect=http_404_error),
+        patch.object(mock_wallbox, "enableEcoSmart", side_effect=http_404_error),
         pytest.raises(HomeAssistantError),
     ):
         await hass.services.async_call(
@@ -129,21 +104,45 @@ async def test_wallbox_select_too_many_requests_error(
     entry: MockConfigEntry,
     mode,
     response,
-    mock_authenticate,
+    mock_wallbox,
 ) -> None:
     """Test wallbox select class connection error."""
 
-    await setup_integration_select(hass, entry, response)
+    await setup_integration(hass, entry)
 
     with (
-        patch(
-            "homeassistant.components.wallbox.Wallbox.disableEcoSmart",
-            new=Mock(side_effect=http_429_error),
-        ),
-        patch(
-            "homeassistant.components.wallbox.Wallbox.enableEcoSmart",
-            new=Mock(side_effect=http_429_error),
-        ),
+        patch.object(mock_wallbox, "getChargerStatus", return_value=response),
+        patch.object(mock_wallbox, "disableEcoSmart", side_effect=http_429_error),
+        patch.object(mock_wallbox, "enableEcoSmart", side_effect=http_429_error),
+        pytest.raises(HomeAssistantError),
+    ):
+        await hass.services.async_call(
+            SELECT_DOMAIN,
+            SERVICE_SELECT_OPTION,
+            {
+                ATTR_ENTITY_ID: MOCK_SELECT_ENTITY_ID,
+                ATTR_OPTION: mode,
+            },
+            blocking=True,
+        )
+
+
+@pytest.mark.parametrize(("mode", "response"), TEST_OPTIONS)
+async def test_wallbox_select_connection_error(
+    hass: HomeAssistant,
+    entry: MockConfigEntry,
+    mode,
+    response,
+    mock_wallbox,
+) -> None:
+    """Test wallbox select class connection error."""
+
+    await setup_integration(hass, entry)
+
+    with (
+        patch.object(mock_wallbox, "getChargerStatus", return_value=response),
+        patch.object(mock_wallbox, "disableEcoSmart", side_effect=ConnectionError),
+        patch.object(mock_wallbox, "enableEcoSmart", side_effect=ConnectionError),
         pytest.raises(HomeAssistantError),
     ):
         await hass.services.async_call(
