+++ tests/components/wallbox/test_switch.py
@@ -1,23 +1,22 @@
 """Test Wallbox Lock component."""
 
-from unittest.mock import Mock, patch
+from unittest.mock import patch
 
 import pytest
 
 from homeassistant.components.switch import SERVICE_TURN_OFF, SERVICE_TURN_ON
-from homeassistant.components.wallbox.const import CHARGER_STATUS_ID_KEY
 from homeassistant.const import ATTR_ENTITY_ID
 from homeassistant.core import HomeAssistant
 from homeassistant.exceptions import HomeAssistantError
 
-from . import authorisation_response, http_404_error, http_429_error, setup_integration
+from .conftest import http_404_error, http_429_error, setup_integration
 from .const import MOCK_SWITCH_ENTITY_ID
 
 from tests.common import MockConfigEntry
 
 
 async def test_wallbox_switch_class(
-    hass: HomeAssistant, entry: MockConfigEntry
+    hass: HomeAssistant, entry: MockConfigEntry, mock_wallbox
 ) -> None:
     """Test wallbox switch class."""
 
@@ -27,55 +26,34 @@ async def test_wallbox_switch_class(
     assert state
     assert state.state == "on"
 
-    with (
-        patch(
-            "homeassistant.components.wallbox.Wallbox.authenticate",
-            new=Mock(return_value=authorisation_response),
-        ),
-        patch(
-            "homeassistant.components.wallbox.Wallbox.pauseChargingSession",
-            new=Mock(return_value={CHARGER_STATUS_ID_KEY: 193}),
-        ),
-        patch(
-            "homeassistant.components.wallbox.Wallbox.resumeChargingSession",
-            new=Mock(return_value={CHARGER_STATUS_ID_KEY: 193}),
-        ),
-    ):
-        await hass.services.async_call(
-            "switch",
-            SERVICE_TURN_ON,
-            {
-                ATTR_ENTITY_ID: MOCK_SWITCH_ENTITY_ID,
-            },
-            blocking=True,
-        )
+    await hass.services.async_call(
+        "switch",
+        SERVICE_TURN_ON,
+        {
+            ATTR_ENTITY_ID: MOCK_SWITCH_ENTITY_ID,
+        },
+        blocking=True,
+    )
 
-        await hass.services.async_call(
-            "switch",
-            SERVICE_TURN_OFF,
-            {
-                ATTR_ENTITY_ID: MOCK_SWITCH_ENTITY_ID,
-            },
-            blocking=True,
-        )
+    await hass.services.async_call(
+        "switch",
+        SERVICE_TURN_OFF,
+        {
+            ATTR_ENTITY_ID: MOCK_SWITCH_ENTITY_ID,
+        },
+        blocking=True,
+    )
 
 
 async def test_wallbox_switch_class_connection_error(
-    hass: HomeAssistant, entry: MockConfigEntry
+    hass: HomeAssistant, entry: MockConfigEntry, mock_wallbox
 ) -> None:
     """Test wallbox switch class connection error."""
 
     await setup_integration(hass, entry)
 
     with (
-        patch(
-            "homeassistant.components.wallbox.Wallbox.authenticate",
-            new=Mock(return_value=authorisation_response),
-        ),
-        patch(
-            "homeassistant.components.wallbox.Wallbox.resumeChargingSession",
-            new=Mock(side_effect=http_404_error),
-        ),
+        patch.object(mock_wallbox, "resumeChargingSession", side_effect=http_404_error),
         pytest.raises(HomeAssistantError),
     ):
         # Test behavior when a connection error occurs
@@ -90,21 +68,14 @@ async def test_wallbox_switch_class_connection_error(
 
 
 async def test_wallbox_switch_class_too_many_requests(
-    hass: HomeAssistant, entry: MockConfigEntry
+    hass: HomeAssistant, entry: MockConfigEntry, mock_wallbox
 ) -> None:
     """Test wallbox switch class connection error."""
 
     await setup_integration(hass, entry)
 
     with (
-        patch(
-            "homeassistant.components.wallbox.Wallbox.authenticate",
-            new=Mock(return_value=authorisation_response),
-        ),
-        patch(
-            "homeassistant.components.wallbox.Wallbox.resumeChargingSession",
-            new=Mock(side_effect=http_429_error),
-        ),
+        patch.object(mock_wallbox, "resumeChargingSession", side_effect=http_429_error),
         pytest.raises(HomeAssistantError),
     ):
         # Test behavior when a connection error occurs
